dom/media/test/test_bug895091.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Bug 895091 - Integrating vtt.js</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="manifest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<video id="v" src="seek.webm" preload="metadata">
  <track src="long.vtt" kind="subtitles" id="track1">
  <track src="long.vtt" kind="subtitles" id="track2">
</video>
<script type="text/javascript">

var trackElement = document.getElementById("track1");
var trackElementTwo = document.getElementById("track2");

async function runTest() {
  enableBothTracks();
  await waitUntilBothTracksLoaded();

  // Set mode to hidden so that the active cue lists are being updated.
  trackElement.track.mode = "hidden";
  trackElementTwo.track.mode = "hidden";

  is(trackElement.readyState, 2, "Track::ReadyState should be set to LOADED.");
  is(trackElementTwo.readyState, 2, "Track::ReadyState should be set to LOADED.");

  is(trackElement.track.cues.length, 2000, "Cue list length should be 2000.");
  is(trackElementTwo.track.cues.length, 2000, "Cue list length should be 2000.");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
onload = runTest;

/**
 * The following are test helper functions.
 */
function enableBothTracks() {
  // All tracks are `disable` on default. As we won't start loading for disabled
  // tracks, we have to change their mode in order to start loading.
  trackElement.track.mode = "hidden";
  trackElementTwo.track.mode = "hidden";
}

async function waitUntilBothTracksLoaded() {
  info(`wait until both tracks finish loading`);
  await Promise.all([once(trackElement, "load"), once(trackElementTwo, "load")]);
}

</script>
</body>
</html>
